add gacha system (random selection)

+++ app/models/Model.scala
@@ -1,4 +1,4 @@
 package models
 import org.joda.time.{DateTime, DateTimeZone}
 
-case class Jenres(id: BigInt, name: String, sentence: String)
+case class Jenres(id: Int, name: String, sentence: String)
